chore: rename and move modal
