+++ web/src/lib/modals/LibraryExclusionPatternModal.svelte
